fixing warnings, refs #3, #2495
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@21210 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSRightOfWayJunction.cpp
@@ -150,7 +150,7 @@ MSRightOfWayJunction::postloadInit() {
 #ifdef HAVE_INTERNAL_LANES
             if (MSGlobals::gUsingInternalLanes && myInternalLanes.size() > 0) {
                 int li = 0;
-                for (int c = 0; c < sortedLinks.size(); ++c) {
+                for (int c = 0; c < (int)sortedLinks.size(); ++c) {
                     if (sortedLinks[c].second->getLane() == 0) { // dead end
                         continue;
                     }
